refactor: restruture c2pcli commands
move viewer back to the top level cmd/ and remove
the tools subcommand
