+++ .secrets.baseline
@@ -3,7 +3,7 @@
     "files": "go.sum|^.secrets.baseline$",
     "lines": null
   },
-  "generated_at": "2025-03-17T22:40:21Z",
+  "generated_at": "2025-03-18T12:24:01Z",
   "plugins_used": [
     {
       "name": "AWSKeyDetector"
@@ -110,7 +110,7 @@
         "hashed_secret": "21c50805b553b7a40e48394a5d77d442587ddee2",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 188,
+        "line_number": 193,
         "type": "Secret Keyword",
         "verified_result": null
       }
