--- cmd/c2pcli/cli/subcommands/tools/cmd.go
@@ -1,33 +1,0 @@
-/*
-Copyright 2023 IBM Corporation
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package tools
-
-import (
-	"github.com/spf13/cobra"
-
-	"github.com/oscal-compass/compliance-to-policy-go/v2/cmd/c2pcli/cli/subcommands/tools/oscal2posture"
-	"github.com/oscal-compass/compliance-to-policy-go/v2/cmd/c2pcli/cli/subcommands/tools/viewer"
-)
-
-func New() *cobra.Command {
-	command := &cobra.Command{
-		Use:   "tools",
-		Short: "Tools for working with OSCAL Documents",
-	}
-	command.AddCommand(oscal2posture.New(), viewer.New())
-	return command
-}
