--- docs/tools/README.md
@@ -1,9 +1,0 @@
-## Tools
-
-### OSCAL2Posture
-
-```bash
-c2pcli tools oscal2posture -c ./docs/ocm/c2p-config.yaml --assessment-results /tmp/assessment-results.json -o /tmp/compliance-posture.md
-```
-
-You can view an example compliance posture like [compliance-posture.md](../ocm/final-outputs/compliance-posture.md)
